+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - <kbd>Ctrl</kbd>+right-click (Windows/Linux) or <kbd>Cmd</kbd>+right-click (macOS) to do the reverse rotation
 - Presets can now be renamed by double-clicking on them
 - Help hover docs for HPS dir and piece filters
+- Warning when launching from a ZIP file
 
 ### Changed